Created Integration Test class, but its not complete yet Refactored Controller Layer Unit Test
